+++ RELEASE-NOTES.md
@@ -4,6 +4,7 @@ Release Notes
 # 0.5.0-SNAPSHOT (in progress)
 - refactored units to be a standalone top-level type
 - added conversion function for unit based values
+- refactored persistence service interface and abstract persistence service base implementation (breaking API changes)
 - ...
 
 # 0.4.0